+++ config/mapprovider/envmapprovider/mapprovider.go
@@ -39,13 +39,13 @@ func New() config.MapProvider {
 
 func (emp *mapProvider) Retrieve(_ context.Context, uri string, _ config.WatcherFunc) (config.Retrieved, error) {
 	if !strings.HasPrefix(uri, schemeName+":") {
-		return config.Retrieved{}, fmt.Errorf("%v uri is not supported by %v provider", uri, schemeName)
+		return nil, fmt.Errorf("%v uri is not supported by %v provider", uri, schemeName)
 	}
 
 	content := os.Getenv(uri[len(schemeName)+1:])
 	var data map[string]interface{}
 	if err := yaml.Unmarshal([]byte(content), &data); err != nil {
-		return config.Retrieved{}, fmt.Errorf("unable to parse yaml: %w", err)
+		return nil, fmt.Errorf("unable to parse yaml: %w", err)
 	}
 
 	return config.NewRetrievedFromMap(config.NewMapFromStringMap(data)), nil