+++ service/mapresolver_test.go
@@ -41,7 +41,7 @@ type mockProvider struct {
 
 func (m *mockProvider) Retrieve(_ context.Context, _ string, watcher config.WatcherFunc) (config.Retrieved, error) {
 	if m.errR != nil {
-		return config.Retrieved{}, m.errR
+		return nil, m.errR
 	}
 	if m.retM == nil {
 		return config.NewRetrievedFromMap(config.NewMap()), nil